Rename use_cases/missions to mission in frontend
