+++ frontend/src/api/APIWorker.tsx
@@ -13,7 +13,7 @@ import getAllGearCodes from '../domain/use_cases/gearCode/getAllGearCodes'
 import getHealthcheck from '../domain/use_cases/healthcheck/getHealthcheck'
 import getFishingInfractions from '../domain/use_cases/infraction/getFishingInfractions'
 import getAllRegulatoryLayers from '../domain/use_cases/layer/regulation/getAllRegulatoryLayers'
-import { getVesselControls } from '../domain/use_cases/missions/getVesselControls'
+import { getVesselControls } from '../domain/use_cases/mission/getVesselControls'
 import { getAllCurrentReportings } from '../domain/use_cases/reporting/getAllCurrentReportings'
 import getAllSpecies from '../domain/use_cases/species/getAllSpecies'
 import getVesselReportings from '../domain/use_cases/vessel/getVesselReportings'
