+++ frontend/src/domain/actions.ts
@@ -1,5 +1,5 @@
 import { missionSliceActions } from './shared_slices/Mission'
-import { missionDispatchers } from './use_cases/missions'
+import { missionDispatchers } from './use_cases/mission'
 
 export const missionActions = {
   ...missionDispatchers,
