+++ frontend/src/domain/use_cases/mission/addControlCoordinates.ts
