+++ frontend/src/domain/use_cases/mission/closeDraw.ts
