+++ frontend/src/domain/use_cases/mission/getVesselControls.ts
