+++ frontend/src/domain/use_cases/mission/index.ts
