+++ frontend/src/domain/use_cases/mission/validateMissionZone.ts
