+++ frontend/src/features/map/draw/DrawModal.tsx
@@ -18,7 +18,7 @@ import { setInteractionType } from '../../../domain/shared_slices/Draw'
 import { fitToExtent } from '../../../domain/shared_slices/Map'
 import { addFeatureToDrawedFeature } from '../../../domain/use_cases/draw/addFeatureToDrawedFeature'
 import { eraseDrawedGeometries } from '../../../domain/use_cases/draw/eraseDrawedGeometries'
-import { closeDraw } from '../../../domain/use_cases/missions/closeDraw'
+import { closeDraw } from '../../../domain/use_cases/mission/closeDraw'
 import { useMainAppDispatch } from '../../../hooks/useMainAppDispatch'
 import { useMainAppSelector } from '../../../hooks/useMainAppSelector'
 import { theme } from '../../../ui/theme'
